requirements.txt
pandas
numpy
patool==4.0.1
playwright==1.52.0
selenium==4.33.0
